+++ app/src/main/java/com/example/studyapp/ui/screens/topics/TopicsPaneContent.kt
@@ -1,31 +1,21 @@
 package com.example.studyapp.ui.screens.topics
 
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.Checkbox
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
-import androidx.compose.material3.ListItem
-import androidx.compose.material3.ListItemColors
-import androidx.compose.material3.ListItemDefaults
 import androidx.compose.material3.LoadingIndicator
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.example.studyapp.R
 import com.example.studyapp.data.Topic
 import com.example.studyapp.ui.components.DockedSearchBar
 import com.example.studyapp.ui.components.PlaceholderColumn
+import com.example.studyapp.ui.components.study.TopicsLazyColumn
 
 @OptIn(ExperimentalMaterial3ExpressiveApi::class)
 @Composable
@@ -55,37 +45,12 @@ fun TopicsPaneContent(
                 modifier = modifier.fillMaxSize()
             )
         } else {
-            ScrollableTopicsList(
+            TopicsLazyColumn(
                 topics = topics,
                 navigateToTopic = navigateToTopic,
                 modifier = modifier
                     .padding(horizontal = 8.dp)
                     .fillMaxSize()
-            )
-        }
-    }
-}
-
-@Composable
-fun ScrollableTopicsList(
-    topics: List<Topic>,
-    navigateToTopic: (Int) -> Unit,
-    modifier: Modifier = Modifier,
-    selectedTopicId: Int? = null,
-) {
-    LazyColumn(modifier = modifier) {
-        items(topics) { topic ->
-            var colors = ListItemDefaults.colors()
-            if (selectedTopicId == topic.id) {
-                colors =
-                    ListItemDefaults.colors(containerColor = MaterialTheme.colorScheme.secondaryContainer)
-            }
-            TopicListItem(
-                topic = topic,
-                colors = colors,
-                modifier = Modifier
-                    .clickable { navigateToTopic(topic.id) }
-                    .fillMaxWidth()
             )
         }
     }
@@ -105,7 +70,7 @@ private fun TopicsSearchBar(
         itemLabel = { it.title },
         placeholderText = stringResource(R.string.search_in_topics)
     ) {
-        ScrollableTopicsList(
+        TopicsLazyColumn(
             topics = it,
             navigateToTopic = navigateToTopic,
             modifier = modifier
@@ -115,29 +80,4 @@ private fun TopicsSearchBar(
     }
 }
 
-@Composable
-private fun TopicListItem(
-    topic: Topic,
-    modifier: Modifier = Modifier,
-    colors: ListItemColors = ListItemDefaults.colors(),
-) {
-    ListItem(
-        headlineContent = {
-            Text(
-                text = topic.title,
-                overflow = TextOverflow.Ellipsis,
-                maxLines = 1
-            )
-        },
-        modifier = modifier,
-        trailingContent = {
-            Checkbox(
-                checked = topic.checked,
-                enabled = false,
-                onCheckedChange = null,
-                modifier = Modifier.size(size = 24.dp)
-            )
-        },
-        colors = colors
-    )
-}
+
